http: do not use local settings for Date header
